More lint cleanup.
In this commit, the following warnings have been dealt with:

- 1506: Call to virtual function 'Symbol' within a constructor or destructor

Many files have been modified here, and for a good reason.  The call to the virtual function destroy in StaticDialog can't be virtual since you can't call a virtual function in a destructor. So I had to make sure that all the destroy functions for the descendants of StaticDialog where correctly implemented, and called from their own destructor.

Doing that, I also realized that most of the classes that derived from Window implemented the exact same thing for their destructor, so I moved that back to Window, and I made sure that all the Window descendants either implemented destroy properly and called it properly, or it they just had the default debavior, I stripped it entirely.

Signed-off-by: Joce <jocelynlegault@gmail.com>