Move FindReplageDlg.h from ScintillaComponent to ScitillaComponent.
Moved to accomodate merge that failed to track the change since the changes
are large compared to the size of the file.
This issue was brought on #git and Ebridius coded a fix for that that would
allow to modify the move treshold at merge time, but it's not yet available
to me.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>