Move Notepad_plus::command to its own file.
In preparation for the next commit to integrate, the Notepad_plus::command
method was moved to its own file. This will facilitate the coming merge.
I also had to make modifications to Notepad_plus.h to remove a the static
qualifier to a function declared in there (doesn't make any sense) that is
used in both Notepad_plus.cpp and NppCommands.cpp.
The modifications in UserDefineDialog.* are somewhat unrelated, but following
the migration of the method, I got a weird compilation error related to static
consts defined in a class. This fix is temporary, I hope.

[#45]

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>