Fix heap corruption with color picker/popup.
ColorPicker owns a colorPopup and maintains its life cycle.
However, when colorPopup sent either a WM_PICKUP_COLOR or a WM_PICKUP_CANCEL
to its parent (the colorPicker), the parent deleted the colorPopup. Therefore
the returning call ended up in a destroyed object.
The fix is rather simple. Instead of constantly creating and deleting the
colorPopup, we simply create it the first time it's used, and hide and show it
afterwards. The colorPopup is finally destroyed when the colorPicker is destroyed.

[#45]

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>