Fix font size bug related to session loading.
The bug:
When loading a session where the sub view contained files but not the main
one, and the active session was the main one. The impact of such a session
file was weird, as the font size used for the sub view was 0 or 1, making the
text illegible.

The repro steps:
http://joce.pastebin.com/fvd7P1kD

The screenshots:
These are screenshots of the same session, one where the files are in the main
view (loads correctly) and one where the files are in the sub view (when the
bug is visible):
When it loads correctly: http://is.gd/g40R5
When the sessions are inverted and the bug shows: http://is.gd/g40WC

The fix:
- factorizes the session loading code in a function.
- loads the main view first
- loads the sub view second, but loads it as the main view if the main view is
  still empty.
- loads the index of the current active view only if the sub view isn't empty
  and sets it to 0 (the index of the main view).

All this as for effect that no files will be loaded in the sub view if the
main view is empty.

[#45]

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>