Really fix the font size issue.
After much investigation, it was found that the bug appeared mainly in debug.
After a few debug session, the issue was traced to an uninitialized zoom
variable that defaulted to 0xcdcdcdcd, or -842150451.
The zoom variable is now properly fixed and the bug can no longer be
reproduced.

The previous "fix" finally wasn't one at all. It just hid the behavior since
the sub view wasn't shown. I do believe however that it's still a good
refactor and a better behavior and should be kept.

[#45]

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>